refactor: remove unused components, blog posts, and outdated utility functions
- Deletes legacy components (`Badge`, `Button`, `Card`, `BackToTop`, `Bluesky`, `GitHub`, `LinkedIn`, `theme-toggle`, and `sidebar`) to simplify the codebase.
- Removes unused blog posts and their metadata to declutter the project.
- Cleans up outdated utility functions, including `fetchMarkdownPosts` from `utils/index.ts`, as they are no longer required.
- Refines project structure for improved maintainability.
