+++ .agents/skills/conventional-commit/SKILL.md
@@ -3,7 +3,7 @@ name: conventional-commit
 description: Create conventional commit messages following best conventions. Use when committing code changes, writing commit messages, or formatting git history. Follows conventional commits specification.
 license: MIT
 metadata:
-  version: "1.0.0"
+  version: '1.0.0'
 ---
 
 # Conventional Commit Messages
@@ -42,20 +42,20 @@ The header is required. Scope is optional. All lines must stay under 100 charact
 
 ## Commit Types
 
-| Type | Purpose |
-|------|---------|
-| `build` | Build system or CI changes |
-| `chore` | Routine maintenance tasks |
-| `ci` | Continuous integration configuration |
-| `deps` | Dependency updates |
-| `docs` | Documentation changes |
-| `feat` | New feature |
-| `fix` | Bug fix |
-| `perf` | Performance improvement |
+| Type       | Purpose                               |
+| ---------- | ------------------------------------- |
+| `build`    | Build system or CI changes            |
+| `chore`    | Routine maintenance tasks             |
+| `ci`       | Continuous integration configuration  |
+| `deps`     | Dependency updates                    |
+| `docs`     | Documentation changes                 |
+| `feat`     | New feature                           |
+| `fix`      | Bug fix                               |
+| `perf`     | Performance improvement               |
 | `refactor` | Code refactoring (no behavior change) |
-| `revert` | Revert a previous commit |
-| `style` | Code style and formatting |
-| `test` | Tests added, updated or improved |
+| `revert`   | Revert a previous commit              |
+| `style`    | Code style and formatting             |
+| `test`     | Tests added, updated or improved      |
 
 ## Subject Line Rules
 
@@ -72,6 +72,7 @@ The header is required. Scope is optional. All lines must stay under 100 charact
 - Contrast with previous behavior when relevant
 
 ## Conventional Commits
+
 The commit contains the following structural elements, to communicate intent to the consumers of your library:
 
 - fix: a commit of the type fix patches a bug in your codebase (this correlates with PATCH in Semantic Versioning).
@@ -79,7 +80,6 @@ The commit contains the following structural elements, to communicate intent to 
 - BREAKING CHANGE: a commit that has a footer BREAKING CHANGE:, or appends a ! after the type/scope, introduces a breaking API change (correlating with MAJOR in Semantic Versioning). A BREAKING CHANGE can be part of commits of any type.
 - types other than fix: and feat: are allowed, for example @commitlint/config-conventional (based on the Angular convention) recommends build:, chore:, ci:, docs:, style:, refactor:, perf:, test:, and others.
 - footers other than BREAKING CHANGE: <description> may be provided and follow a convention similar to git trailer format.
-
 
 ## Examples
