+++ .prettierignore
@@ -11,3 +11,6 @@ node_modules
 pnpm-lock.yaml
 package-lock.json
 yarn.lock
+
+# prettier-plugin-svelte cannot parse <script type="application/ld+json"> with Svelte expressions
+src/lib/components/seo.svelte
