+++ package.json
@@ -20,16 +20,13 @@
 		"@typescript-eslint/eslint-plugin": "^8.56.0",
 		"@typescript-eslint/parser": "^8.56.0",
 		"autoprefixer": "^10.4.24",
-		"dayjs": "^1.11.19",
 		"eslint": "^10.0.1",
 		"eslint-config-prettier": "^10.1.8",
 		"eslint-plugin-svelte": "^3.15.0",
 		"globals": "^17.3.0",
-		"mdsvex": "^0.12.6",
-		"mdsvex-relative-images": "^2.0.0",
 		"postcss": "^8.5.6",
 		"prettier": "^3.8.1",
-		"prettier-plugin-svelte": "^3.5.0",
+		"prettier-plugin-svelte": "^3.4.1",
 		"prettier-plugin-tailwindcss": "^0.7.2",
 		"svelte": "^5.53.2",
 		"svelte-check": "^4.4.3",
