--- src/app.css
@@ -1,130 +1,0 @@
-@import 'tailwindcss';
-
-@theme {
-	--font-beanie: 'Reenie Beanie';
-
-	/* Design System - Custom Properties */
-	--color-primary: #3b82f6;
-	--color-primary-dark: #2563eb;
-	--color-accent: #f59e0b;
-	--color-accent-hover: #d97706;
-
-	/* Gradients */
-	--gradient-hero: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
-	--gradient-accent: linear-gradient(135deg, #f59e0b 0%, #ef4444 100%);
-
-	/* Prose colors */
-	--color-prose-body: theme(--color-blue-800);
-	--color-prose-headings: theme(--color-blue-950);
-	--color-prose-lead: theme(--color-blue-700);
-	--color-prose-links: theme(--color-blue-950);
-	--color-prose-bold: theme(--color-blue-950);
-	--color-prose-counters: theme(--color-blue-600);
-	--color-prose-bullets: theme(--color-blue-400);
-	--color-prose-hr: theme(--color-blue-300);
-	--color-prose-quotes: theme(--color-blue-950);
-	--color-prose-quote-borders: theme(--color-blue-300);
-	--color-prose-captions: theme(--color-blue-700);
-	--color-prose-code: theme(--color-blue-950);
-	--color-prose-pre-code: theme(--color-blue-100);
-	--color-prose-pre-bg: theme(--color-blue-950);
-	--color-prose-th-borders: theme(--color-blue-300);
-	--color-prose-td-borders: theme(--color-blue-200);
-	--color-prose-invert-body: theme(--color-blue-200);
-	--color-prose-invert-headings: theme(--color-blue-50);
-	--color-prose-invert-lead: theme(--color-blue-300);
-	--color-prose-invert-links: theme(--color-blue-50);
-	--color-prose-invert-bold: theme(--color-blue-50);
-	--color-prose-invert-counters: theme(--color-blue-400);
-	--color-prose-invert-bullets: theme(--color-blue-600);
-	--color-prose-invert-hr: theme(--color-blue-700);
-	--color-prose-invert-quotes: theme(--color-blue-100);
-	--color-prose-invert-quote-borders: theme(--color-blue-700);
-	--color-prose-invert-captions: theme(--color-blue-400);
-	--color-prose-invert-code: theme(--color-blue-50);
-	--color-prose-invert-pre-code: theme(--color-blue-300);
-	--color-prose-invert-pre-bg: rgb(239 246 255 / 20%);
-	--color-prose-invert-th-borders: theme(--color-blue-600);
-	--color-prose-invert-td-borders: theme(--color-blue-700);
-}
-
-@layer base {
-	body {
-		@apply flex min-h-screen flex-col justify-between scroll-smooth bg-blue-50 text-blue-800 antialiased dark:bg-stone-950 dark:text-blue-200;
-	}
-}
-
-@utility animation-delay-2000 {
-	animation-delay: 2s;
-}
-
-@utility animation-delay-4000 {
-	animation-delay: 4s;
-}
-
-/* Custom Animations */
-@keyframes fade-in-up {
-	from {
-		opacity: 0;
-		transform: translateY(20px);
-	}
-	to {
-		opacity: 1;
-		transform: translateY(0);
-	}
-}
-
-@keyframes fade-in {
-	from {
-		opacity: 0;
-	}
-	to {
-		opacity: 1;
-	}
-}
-
-@keyframes slide-in-left {
-	from {
-		opacity: 0;
-		transform: translateX(-20px);
-	}
-	to {
-		opacity: 1;
-		transform: translateX(0);
-	}
-}
-
-@keyframes typing {
-	from {
-		width: 0;
-	}
-	to {
-		width: 100%;
-	}
-}
-
-@keyframes gradient-shift {
-	0%, 100% {
-		background-position: 0% 50%;
-	}
-	50% {
-		background-position: 100% 50%;
-	}
-}
-
-.animate-fade-in-up {
-	animation: fade-in-up 0.6s ease-out;
-}
-
-.animate-fade-in {
-	animation: fade-in 0.6s ease-out;
-}
-
-.animate-slide-in-left {
-	animation: slide-in-left 0.6s ease-out;
-}
-
-.animate-gradient {
-	background-size: 200% 200%;
-	animation: gradient-shift 3s ease infinite;
-}
