+++ src/app.html
@@ -7,26 +7,6 @@
 		<link rel="dns-prefetch" href="https://fonts.googleapis.com" />
 		<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
 		<link
-			rel="preload"
-			as="style"
-			href="https://fonts.googleapis.com/css?family=Reenie+Beanie&display=swap&text=DanielSalvado"
-		/>
-		<link
-			rel="stylesheet"
-			media="print"
-			onload="
-				this.onload = null;
-				this.removeAttribute('media');
-			"
-			href="https://fonts.googleapis.com/css?family=Reenie+Beanie&display=swap&text=DanielSalvado"
-		/>
-		<noscript>
-			<link
-				rel="stylesheet"
-				href="https://fonts.googleapis.com/css?family=Reenie+Beanie&display=swap&text=DanielSalvado"
-			/>
-		</noscript>
-		<link
 			rel="apple-touch-icon-precomposed"
 			sizes="57x57"
 			href="%sveltekit.assets%/favicon/apple-touch-icon-57x57.png"
