--- src/lib/components/BackToTop.svelte
@@ -1,40 +1,0 @@
-<script lang="ts">
-	import { onMount } from 'svelte';
-
-	let visible = false;
-
-	onMount(() => {
-		const handleScroll = () => {
-			visible = window.scrollY > 300;
-		};
-
-		window.addEventListener('scroll', handleScroll);
-		return () => window.removeEventListener('scroll', handleScroll);
-	});
-
-	function scrollToTop() {
-		window.scrollTo({ top: 0, behavior: 'smooth' });
-	}
-</script>
-
-{#if visible}
-	<button
-		on:click={scrollToTop}
-		class="fixed bottom-8 right-8 z-30 p-3 rounded-full bg-blue-600 hover:bg-blue-700 text-white shadow-lg hover:shadow-xl transition-all duration-300 hover:scale-110 focus:outline-none focus:ring-2 focus:ring-blue-500 focus:ring-offset-2 animate-fade-in"
-		aria-label="Back to top"
-	>
-		<svg
-			class="w-6 h-6"
-			fill="none"
-			stroke="currentColor"
-			viewBox="0 0 24 24"
-		>
-			<path
-				stroke-linecap="round"
-				stroke-linejoin="round"
-				stroke-width="2"
-				d="M5 10l7-7m0 0l7 7m-7-7v18"
-			/>
-		</svg>
-	</button>
-{/if}
