+++ src/lib/components/Footer.svelte
@@ -1,0 +1,14 @@
+<script lang="ts">
+	const year = new Date().getFullYear();
+</script>
+
+<footer>
+	<div class="flex flex-col items-center justify-between gap-6 sm:flex-row">
+		<p>
+			&copy; {year} Daniel Salvado. All rights reserved.
+		</p>
+		<a href="/legal-notice">
+			Legal Notice
+		</a>
+	</div>
+</footer>
