--- src/lib/components/Twitter.svelte
@@ -1,21 +1,0 @@
-<script lang="ts">
-	const props = $props();
-</script>
-
-<!--iconoir-->
-<svg
-	{...props}
-	viewBox="0 0 24 24"
-	stroke-width="1.5"
-	fill="none"
-	xmlns="http://www.w3.org/2000/svg"
-	color="currentcolor"
-	><path
-		d="M16.8198 20.7684L3.75317 3.96836C3.44664 3.57425 3.72749 3 4.22678 3H6.70655C6.8917 3 7.06649 3.08548 7.18016 3.23164L20.2468 20.0316C20.5534 20.4258 20.2725 21 19.7732 21H17.2935C17.1083 21 16.9335 20.9145 16.8198 20.7684Z"
-		stroke="currentcolor"
-		stroke-width="1.5"
-	></path><path d="M20 3L4 21" stroke="currentcolor" stroke-width="1.5" stroke-linecap="round"></path></svg
->
-
-<!--svgl.app-->
-<!--<svg {...props} fill="none" viewBox="0 0 1200 1227" xmlns="http://www.w3.org/2000/svg"><path d="M714.163 519.284 1160.89 0h-105.86L667.137 450.887 357.328 0H0l468.492 681.821L0 1226.37h105.866l409.625-476.152 327.181 476.152H1200L714.137 519.284h.026ZM569.165 687.828l-47.468-67.894-377.686-540.24h162.604l304.797 435.991 47.468 67.894 396.2 566.721H892.476L569.165 687.854v-.026Z" fill="#fff" /></svg>-->
