+++ src/lib/constants.ts
@@ -1,0 +1,13 @@
+export const BASE_URL = 'https://danielsalvado.com';
+
+export const X_PROFILE_LINK = 'https://x.com/danisalTweets';
+export const BLUESKY_PROFILE_URL = 'https://bsky.app/profile/danisal.bsky.social';
+export const LINKEDIN_PROFILE_LINK = 'https://www.linkedin.com/in/daniel-salvado/';
+export const GITHUB_PROFILE_URL = 'https://github.com/danisal';
+
+export const socialLinks = {
+	github: { href: GITHUB_PROFILE_URL, label: 'Daniel on GitHub' },
+	linkedin: { href: LINKEDIN_PROFILE_LINK, label: 'Daniel on LinkedIn' },
+	x: { href: X_PROFILE_LINK, label: 'Daniel on X' },
+	bluesky: { href: BLUESKY_PROFILE_URL, label: 'Daniel on Bluesky' }
+};
