+++ src/routes/+layout.svelte
@@ -1,35 +1,36 @@
 <script lang="ts">
 	import SEO from '$lib/components/SEO.svelte';
-	import Footer from '$lib/components/footer.svelte';
-	import Header from '$lib/components/header.svelte';
-	import Sidebar from '$lib/components/sidebar.svelte';
-	import BackToTop from '$lib/components/BackToTop.svelte';
-
-	import '../app.css';
-	import { page } from '$app/stores';
-
-	// Control the sidebar on mobile
-	let open: boolean = false;
+	import Footer from '$lib/components/Footer.svelte';
+	import '../styles.css';
+	import { page } from '$app/state';
+	import {
+		BASE_URL,
+		BLUESKY_PROFILE_URL,
+		GITHUB_PROFILE_URL,
+		LINKEDIN_PROFILE_LINK,
+		X_PROFILE_LINK
+	} from '$lib/constants';
+	import Header from '$lib/components/Header.svelte';
 
 	const meta = {
 		title: 'Home',
 		description: 'Daniel Salvado Virtual Space',
-		url: 'https://danielsalvado.com'
+		url: BASE_URL
 	};
 
 	// SEO
 	$: seoProps = {
-		title: $page.data.seoTitle || meta.title,
+		title: page.data.seoTitle || meta.title,
 		titleTemplate: '%s | Daniel Salvado',
-		description: $page.data.seoDescription || meta.description,
-		keywords: $page.data.seoKeywords || [],
+		description: page.data.seoDescription || meta.description,
+		keywords: page.data.seoKeywords || [],
 		openGraph: {
 			type: 'website',
 			url: meta.url,
 			description: meta.description,
 			images: [
 				{
-					url: 'https://danielsalvado.com/images/1200x630.png',
+					url: `${BASE_URL}/images/1200x630.png`,
 					alt: 'Daniel Salvado Banner',
 					width: 1200,
 					height: 675
@@ -38,9 +39,9 @@
 		},
 		twitter: {
 			cardType: 'summary_large_image',
-			title: $page.data.seoTitle || meta.title,
+			title: page.data.seoTitle || meta.title,
 			description: meta.description,
-			image: 'https://danielsalvado.com/images/1200x630.png',
+			image: `${BASE_URL}/images/1200x630.png`,
 			imageAlt: 'Daniel Salvado Banner',
 			site: meta.url
 		},
@@ -49,12 +50,13 @@
 			'@context': 'https://schema.org',
 			'@type': 'WebSite',
 			name: 'Daniel Salvado',
-			url: 'https://danielsalvado.com',
+			url: BASE_URL,
 			description: 'Daniel Salvado personal website',
 			sameAs: [
-				'https://www.linkedin.com/in/daniel-salvado/',
-				'https://twitter.com/danisalTweets',
-				'https://github.com/danisal'
+				LINKEDIN_PROFILE_LINK,
+				X_PROFILE_LINK,
+				BLUESKY_PROFILE_URL,
+				GITHUB_PROFILE_URL,
 			]
 		}
 	};
@@ -63,8 +65,6 @@
 <!-- SEO -->
 <SEO {...seoProps} />
 
-<Sidebar bind:open />
-<Header bind:open />
+<Header />
 <slot />
 <Footer />
-<BackToTop />
