+++ src/routes/blog/+page.server.ts
@@ -1,9 +1,5 @@
-export const load = async ({ fetch }) => {
-	const response = await fetch(`/api/posts`);
-	const posts = await response.json();
+import { redirect } from '@sveltejs/kit';
 
-	return {
-		posts,
-		seoTitle: 'All Post'
-	};
+export const load = async () => {
+	throw redirect(301, '/');
 };
