+++ src/routes/blog/[slug]/+page.server.ts
@@ -1,23 +1,5 @@
-import type { PageServerLoad } from './$types';
+import { redirect } from '@sveltejs/kit';
 
-type Post = {
-	title: string;
-	date: string;
-	keywords: string[];
+export const load = async () => {
+	throw redirect(301, '/');
 };
-
-export const load = (async ({ params }) => {
-	const post = await import(`../${params.slug}/index.md`);
-	const content = post.default.render();
-
-	const { title, date, keywords }: Post = post.metadata;
-
-	return {
-		content,
-		date,
-		keywords,
-		title,
-		seoTitle: title,
-		seoKeywords: keywords
-	};
-}) satisfies PageServerLoad;
