--- src/routes/blog/hello-world/index.md
@@ -1,53 +1,0 @@
----
-title: Hello World
-date: '2019-07-07'
-keywords:
-  [
-    'new website',
-    'new blog',
-    'technologies',
-    'gatsbyjs',
-    'emotionjs',
-    'react',
-    'seo',
-    'pwa',
-    'progressive web app',
-    'learning'
-  ]
-image: './photo-by-deanna-alys-unsplash.jpeg'
-reading: 3
----
-
-[![Photo by deanna alys on unsplash](./photo-by-deanna-alys-unsplash.jpeg)](https://unsplash.com/photos/ocXWR-Tz8qY)
-
-Hello World from danielsalvado.com!
-
-Finally, my personal website saw the light of day and reached the World Wide Web!
-
-It's a great feeling to have finally deployed my website. It took me some time, mainly because as creators we are never satisfied with our work, or at least that's how it is with me.
-
-Although I have this feeling, that the site could be better, I really wanted to put it online, in order to force me to have a deadline and comply with it. This way there is no going back and I can iterate on it to make improvements and implement new features. This way I can have feedback, suggestions and improve on things that may have escaped me or that are not so obvious to me.
-
-What is the reason for creating my personal website, you may ask?
-
-Well, I can say that the main reason was to get out of the comfort zone, create something from scratch, work with new technologies and improve as a developer.
-
-Another reason is that I'd like to have a space where I can share articles or experiences that I went through as a developer, without having to rely on external platforms or being interrupted all the time when I read articles!
-
-### Regarding the technologies I used and why I used them
-
-I chose to create the site in a static way and use [GatsbyJS](https://www.gatsbyjs.org/) for that, behind the scenes, for those who don't know, Gatsby uses [React](https://reactjs.org/), something I'm familiar with. Also it this project allowed me to work with other frameworks, such as [EmotionJS](https://emotion.sh/docs/introduction) for styling since I've been reading good reviews and I'd like to experiment.
-
-Another great advantage of opting for Gatsby is that it provides a lot of advantages out of the box and besides, it has a great library of plugins if you want to add some more functionality, which is great!
-
-Something I've always liked to keep in mind too, is the user experience, I wanted it to be good and many times I'd ask myself the question, if I were the one using the site what would I like to have?
-
-My initial list of features to have in order to go online was to have a homepage, a page to show some of the work I've already done, my professional experience, a blog and a contact page. So this was the minimal viable product that would be needed to put the site online.
-
-In terms of functionality, I wanted the site to be responsive, have a good browsing experience, be fast (I don't like to wait too long to see/interact with a site), have a good SEO and accessibility score.
-
-Having posted these conditions and now that the site is online, I think I've achieved these goals. Of course, there is always room for improvement and new features to implement, but as I said before, this is an ongoing project which I will iterate on, this way the project won't be simply put on the shelf and I'll be able to keep it up to date in regards to current technologies. And on top of that, I'll be able to improve as a professional doing something I like!
-
-Well, I think that's all for now, I hope you like the site and I'll be grateful for all the feedback and/or suggestions you may have.
-
-See you in the next blog post!
