+++ src/routes/work/+page.server.ts
@@ -1,7 +1,5 @@
+import { redirect } from '@sveltejs/kit';
+
 export const load = async () => {
-	return {
-		seoTitle: 'Work',
-		seoDescription:
-			"Explore Daniel's Work Experience | Dive into my professional journey, accomplishments, and expertise in Computer Science. Discover my career highlights and contributions."
-	};
+	throw redirect(301, '/');
 };
