+++ src/styles.css
@@ -1,0 +1,28 @@
+@import url('https://fonts.googleapis.com/css2?family=Reenie+Beanie&display=swap&text=DanielSalvado');
+@import 'tailwindcss';
+
+@plugin '@tailwindcss/typography';
+
+@theme inline {
+    --font-beanie: 'Reenie Beanie';
+}
+
+@layer base {
+    html {
+        font-size: 18px;
+    }
+
+    body {
+        @apply bg-neutral-50 text-neutral-950 dark:bg-neutral-950 dark:text-neutral-50 container prose dark:prose-invert mx-auto px-6 grid grid-rows-[auto,1fr,auto] h-svh scroll-smooth antialiased transition-colors;
+    }
+
+    a {
+        @apply text-amber-500 dark:text-amber-300 no-underline hover:underline hover:decoration-wavy hover:underline-offset-4 transition-colors;
+    }
+
+    header,
+    main,
+    footer{
+        @apply py-10;
+    }
+}
