+++ static/sitemap.xml
@@ -4,24 +4,6 @@
         <loc>http://danielsalvado.com/</loc>
     </url>
     <url>
-        <loc>http://danielsalvado.com/work</loc>
-    </url>
-    <url>
-        <loc>http://danielsalvado.com/blog</loc>
-    </url>
-    <url>
         <loc>http://danielsalvado.com/legal-notice</loc>
-    </url>
-    <url>
-        <loc>http://danielsalvado.com/blog/responsive-columns-with-css-grid</loc>
-    </url>
-    <url>
-        <loc>http://danielsalvado.com/blog/hello-world</loc>
-    </url>
-    <url>
-        <loc>http://danielsalvado.com/blog/responsive-data-tables-with-css-grid</loc>
-    </url>
-    <url>
-        <loc>http://danielsalvado.com/blog/from-emt-to-web-developer</loc>
     </url>
 </urlset>
