+++ svelte.config.js
@@ -1,16 +1,10 @@
 import adapter from '@sveltejs/adapter-cloudflare';
-import { mdsvex } from 'mdsvex';
-import relativeImages from 'mdsvex-relative-images';
 import { vitePreprocess } from '@sveltejs/vite-plugin-svelte';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	preprocess: [
 		vitePreprocess(),
-		mdsvex({
-			extensions: ['.md'],
-			remarkPlugins: [relativeImages]
-		})
 	],
 	extensions: ['.svelte', '.md'],
 	kit: {
